remove delete button valid passholder, tenant account and subconstructor

+++ resources/views/partials/tenant_detail.blade.php
@@ -23,7 +23,7 @@
 																	<span class="info-box-text">Email: {{ $account->email }}</span>
 																	<span class="text-right info-box-text">
 																		<a href="{{ route('admin.tenant.account.2fa', [$entry->id, $account->id]) }}" class="btn btn-xs btn-default"><i class="fa fa-google-plus-square"></i> Config 2FA</a>
-																		@if($account->id !== auth()->user()->id)
+																		@if($account->id !== auth()->user()->id && !backpack_user()->hasAnyRole([COMPANY_CO_ROLE, COMPANY_AS_ROLE]))
 																			<a href="javascript:void(0)" onclick="deleteEntry(this)" data-route="{{ route('crud.user.destroy', [$account->id]) }}" class="btn btn-xs btn-danger" data-button-type="delete"><i class="fa fa-trash"></i> Delete</a>
 																		@endif
 																	</span>
@@ -52,7 +52,7 @@
 													<span class="info-box-text">Contact: {{ $account->phone }}</span>
 													<span class="info-box-text">Email: {{ $account->email }}</span>
 													<span class="text-right info-box-text">
-														@if($account->id !== auth()->user()->id)
+														@if($account->id !== auth()->user()->id && !backpack_user()->hasAnyRole([COMPANY_CO_ROLE, COMPANY_AS_ROLE]))
 															<a href="javascript:void(0)" onclick="deleteEntry(this)" data-route="{{ route('crud.user.destroy', [$account->id]) }}" class="btn btn-xs btn-danger" data-button-type="delete"><i class="fa fa-trash"></i> Delete</a>
 														@endif
 													</span>
